Split integration tests in two: normal and language support
