+++ .github/workflows/pull_request_integration_tests_lang.yml
@@ -1,0 +1,43 @@
+name: Pull request integration tests for language support
+
+on:
+  push:
+    branches: [ 'main', 'release-*' ]
+  pull_request:
+    branches: [ 'main', 'release-*' ]
+
+jobs:
+  test:
+    name: test
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        go: [ '1.23' ]
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Go
+        uses: actions/setup-go@v3
+        with:
+          go-version: ${{ matrix.go }}
+      - name: Clean up disk space
+        run: |
+          docker system prune -af
+          docker volume prune -f
+      - name: Run integration tests for languages
+        run: make integration-test-lang
+        timeout-minutes: 60
+      - name: Upload integration test logs
+        uses: actions/upload-artifact@v3
+        if: always()
+        with:
+          name: Language test Logs
+          path: |
+            testoutput/*.log
+            testoutput/kind
+      - name: Report coverage
+        uses: codecov/codecov-action@v4
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+        with:
+          file: ./testoutput/itest-covdata.txt
+          flags: lang-integration-test
